+++ src/components/edit-and-share/edit-audio/changeAudio/index.js
@@ -1,32 +1,21 @@
 import React from 'react';
-import { Upload, Button } from 'antd';
-import { UploadOutlined } from '@ant-design/icons';
+import {  Button } from 'antd';
+import { CloseOutlined, PaperClipOutlined  } from '@ant-design/icons';
+import UploadAudio from './UploadAudio';
 import 'antd/dist/antd.css';
 
 class ChangeAudio extends React.Component {
     state = {
+      change: false
     };
-  
-handleChange = info => {
-      let fileList = [...info.fileList];
-      fileList = fileList.slice(-1);
-      fileList = fileList.map(file => {
-        if (file.response) {file.url = file.response.url; }
-        return file;
-      });
-     this.setState({ fileList });
-    };
-  
-    render() {
-      const props = {
-        action: 'https://www.mocky.io/v2/5cc8019d300000980a055e76',
-        onChange: this.handleChange,
-        multiple: true
-      };
-      return (
-        <Upload {...props} fileList={this.state.fileList}>
-              {this.info}
-        </Upload>
+   handleChange = () => this.setState({change : true});
+  render() {
+     return (
+       <div >
+         <PaperClipOutlined /> Url-Audio
+         <Button type="link" handleChange={this.handleChange}><CloseOutlined/></Button>
+         {this.state.change === true && <UploadAudio change={this.state.change}/>}
+         </div>
       );
     }
   }
