+++ src/components/edit-and-share/edit-audio/changeCategory/index.js
@@ -2,16 +2,13 @@ import React from 'react'
 import { Select } from 'antd';
 import 'antd/dist/antd.css';
 
-
-
 function InputChange() {
     const Option  = Select;
     const children = [];
 for (let i = 10; i < 36; i++) {
   children.push(<Option key={i.toString(36) + i}>{i.toString(36) + i}</Option>);
 }
-      
-return(
+ return(
     <Select
     mode="multiple"
     style={{ width: '100%' }}
@@ -21,8 +18,8 @@ return(
     listHeight={250} 
     zIndex={4}>
     {children}
-  </Select>
-);
+    </Select>
+  );
 };
 
 export default InputChange;
