+++ src/components/edit-and-share/index.js
@@ -13,7 +13,7 @@ class EditAndShare extends React.Component {
     
     return(
         <Dropdown overlay={menu} zIndex={2} >
-        <button className="ant-dropdown-link ellipsis" onClick={e => e.preventDefault()}>
+        <button className="ant-dropdown-link" type="link" onClick={e => e.preventDefault()}>
         <EllipsisOutlined/>
         </button>
         </Dropdown>
